requirements.txt
Django==5.1.5
psycopg2-binary==2.9.9
anthropic>=0.31.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
twilio==8.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
